[Refactor] fix DB table
