+++ src/main/java/careerfestival/career/comments/Service/CommentService.java
@@ -2,7 +2,7 @@ package careerfestival.career.comments.Service;
 
 import careerfestival.career.comments.dto.CommentRequestDto;
 import careerfestival.career.comments.dto.CommentResponseDto;
-import careerfestival.career.domain.Comment;
+import careerfestival.career.domain.mapping.Comment;
 import careerfestival.career.domain.Event;
 import careerfestival.career.domain.User;
 import careerfestival.career.repository.CommentRepository;
