--- src/main/java/careerfestival/career/domain/Category.java
@@ -1,33 +1,0 @@
-package careerfestival.career.domain;
-
-import careerfestival.career.domain.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Category extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 200, name = "category_name")
-    private String categoryName;
-
-    @Column(length = 200, name = "category_detail")
-    private String categoryDetail;
-
-    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
-    private List<Record> record = new ArrayList<>();
-
-    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
-    private List<Event> event = new ArrayList<>();
-
-}
