+++ src/main/java/careerfestival/career/domain/Record.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import careerfestival.career.domain.mapping.RecordKeyword;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,6 +19,14 @@ public class Record extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    // category enums
+    @Enumerated(EnumType.STRING)
+    private String category;
+
+    // keyword enums
+    @Enumerated(EnumType.STRING)
+    private String keyword;
+
     @Column(nullable = false, length = 20, name = "event_name")
     private String eventName;
     @Column(nullable = false, name = "event_date")
@@ -32,11 +39,6 @@ public class Record extends BaseEntity {
     @Column(length = 300, name = "networking_contact")
     private String networkingContact;
 
-    @OneToMany(mappedBy = "record", cascade = CascadeType.ALL)
-    private List<RecordKeyword> recordKeywords = new ArrayList<>();
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Category category;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
