+++ src/main/java/careerfestival/career/domain/enums/Category.java
@@ -1,0 +1,5 @@
+package careerfestival.career.domain.enums;
+
+public enum Category {
+    강연, 학술대회, 전시, 기타
+}
