+++ src/main/java/careerfestival/career/domain/mapping/Comment.java
@@ -1,5 +1,7 @@
-package careerfestival.career.domain;
+package careerfestival.career.domain.mapping;
 
+import careerfestival.career.domain.Event;
+import careerfestival.career.domain.User;
 import careerfestival.career.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
