--- src/main/java/careerfestival/career/domain/mapping/RecordKeyword.java
@@ -1,32 +1,0 @@
-package careerfestival.career.domain.mapping;
-
-import careerfestival.career.domain.Record;
-import careerfestival.career.domain.User;
-import careerfestival.career.domain.common.BaseEntity;
-import careerfestival.career.domain.enums.KeywordName;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RecordKeyword extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private KeywordName keywordName;
-
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "record_id")
-    private Record record;
-
-
-}
