+++ src/main/java/careerfestival/career/record/dto/RecordLectureSeminarDto.java
