--- src/main/java/careerfestival/career/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package careerfestival.career.repository;
-
-import careerfestival.career.domain.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
-}
