+++ src/main/java/careerfestival/career/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package careerfestival.career.repository;
 
-import careerfestival.career.domain.Comment;
+import careerfestival.career.domain.mapping.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
