--- src/main/java/careerfestival/career/repository/RecordKeywordRepository.java
@@ -1,8 +1,0 @@
-package careerfestival.career.repository;
-
-import careerfestival.career.domain.mapping.RecordKeyword;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecordKeywordRepository extends JpaRepository<RecordKeyword, Long> {
-
-}
